+++ packages/js/src/initializers/elementor-editor-integration.js
@@ -1,17 +1,11 @@
 import { dispatch } from "@wordpress/data";
 import domReady from "@wordpress/dom-ready";
-import { createPortal, useState } from "@wordpress/element";
 import { doAction } from "@wordpress/hooks";
 import { __, sprintf } from "@wordpress/i18n";
-import { Root } from "@yoast/externals/contexts";
 import { debounce } from "lodash";
-import ElementorSlot from "../elementor/components/slots/ElementorSlot";
-import ElementorFill from "../elementor/containers/ElementorFill";
+import { renderYoastReactRoot, renderYoastTabReactContent } from "../elementor/initializers/render-sidebar";
 import { registerElementorDataHookAfter } from "../helpers/elementorHook";
-import { registerReactComponent, renderReactRoot } from "../helpers/reactRoot";
-import { useMutationObserver } from "../hooks/use-mutation-observer";
-
-const TAB_ID = "yoast-elementor-react-tab";
+import { registerReactComponent } from "../helpers/reactRoot";
 
 // Keep track of unsaved SEO setting changes.
 let hasUnsavedSeoChanges = false;
@@ -224,78 +218,6 @@ function sendFormData( form ) {
 		debouncedUpdateSaveAsDraftWarning();
 	} );
 }
-
-/**
- * Renders the content in a portal if the element exists.
- *
- * The portal is created once the element is detected.
- * The portal is removed once the element is removed.
- *
- * @param {string} id The ID of the element to render in.
- * @param {JSX.node} children The content.
- *
- * @returns {JSX.node|null} The rendered content or null.
- */
-const RenderInPortalIfElementExists = ( { id, children } ) => {
-	const [ render, setRender ] = useState( null );
-
-	useMutationObserver( document.body, () => {
-		const el = document.getElementById( id );
-		if ( el ) {
-			if ( render === null ) {
-				setRender( createPortal( children, el ) );
-			}
-		} else if ( render !== null ) {
-			setRender( null );
-		}
-	} );
-
-	return render;
-};
-
-/**
- * Renders the Yoast tab React content.
- * @returns {void}
- */
-const renderYoastTabReactContent = () => {
-	// Get the current tab/controls.
-	const root = document.getElementById( "elementor-panel-page-settings-controls" );
-	if ( ! root ) {
-		return;
-	}
-
-	// Hide the Elementor control, we just fill the full contents of the tab.
-	const control = root.getElementsByClassName( "elementor-control" )?.[ 0 ];
-	if ( control ) {
-		control.style.display = "none";
-	}
-
-	// Create our Yoast tab inside, being picked up by the MutationObserver of RenderInPortalIfElementExists.
-	const element = document.createElement( "div" );
-	element.id = TAB_ID;
-	element.className = "yoast yoast-elementor-panel__fills";
-	root.appendChild( element );
-};
-
-/**
- * Renders the Yoast React root.
- * @returns {void}
- */
-const renderYoastReactRoot = () => {
-	const elementorSidebarContext = { locationContext: "elementor-sidebar" };
-	const root = document.createElement( "div" );
-	root.id = "yoast-elementor-react-root";
-	document.body.appendChild( root );
-
-	renderReactRoot( root.id, (
-		<Root context={ elementorSidebarContext }>
-			<RenderInPortalIfElementExists id={ TAB_ID }>
-				<ElementorSlot />
-				<ElementorFill />
-			</RenderInPortalIfElementExists>
-		</Root>
-	) );
-};
 
 /**
  * Initializes the Yoast elementor editor integration.
